Improve dev build incrementality
